+++ Tests/UIKitTests/UIEdgeInsetsExtensionsTests.swift
@@ -158,5 +158,19 @@ final class UIEdgeInsetsExtensionsTests: XCTestCase {
 		XCTAssertEqual(negativeComposedInset.top, 10)
 		XCTAssertEqual(negativeComposedInset.bottom, 5.0)
 	}
+
+    func testAddition() {
+        XCTAssertEqual(UIEdgeInsets.zero + UIEdgeInsets.zero, UIEdgeInsets.zero)
+
+        let inset1 = UIEdgeInsets(top: 1, left: 2, bottom: 3, right: 4)
+        let inset2 = UIEdgeInsets(top: 5, left: 6, bottom: 7, right: 8)
+        let expected = UIEdgeInsets(top: 6, left: 8, bottom: 10, right: 12)
+        XCTAssertEqual(inset1 + inset2, expected)
+
+        let negativeInset1 = UIEdgeInsets(top: -1, left: -2, bottom: -3, right: -4)
+        let negativeInset2 = UIEdgeInsets(top: -5, left: -6, bottom: -7, right: -8)
+        let negativeExpected = UIEdgeInsets(top: -6, left: -8, bottom: -10, right: -12)
+        XCTAssertEqual(negativeInset1 + negativeInset2, negativeExpected)
+    }
 }
 #endif